Various fixes and support for Windows Servers using Chef.

+++ lib/dialog/EC2_ImageEBSDeleteDialog.rb
@@ -22,37 +22,44 @@ class EC2_ImageEBSDeleteDialog < FXDialogBox
     @s3_bucket = nil
     @manifest_file = ""
     @file_folder = ""
-    snapshot_id = ""
+    snapshot_ids = Array.new
+    snapshots = ""
     @deleted = false
     r = get_image(sel_image)
-    if r[:root_device_name] != nil
-       root_device = r[:root_device_name]
-       snapshot_id = ""
-       if r[:block_device_mappings] != nil 
-	   r[:block_device_mappings].each do |m|
-	      if m[:device_name] == root_device
-	         snapshot_id = m[:ebs_snapshot_id]	
-	      end
-	   end   
+    begin 
+       if r[:root_device_name] != nil
+          root_device = r[:root_device_name]
+          snapshot_id = ""
+          if r[:block_device_mappings] != nil 
+	     r[:block_device_mappings].each do |m|
+	        puts "s #{m[:ebs_snapshot_id]}"
+                snapshot_ids.push(m[:ebs_snapshot_id])
+                if snapshots == nil or snapshots == ""
+                   snapshots = m[:ebs_snapshot_id]
+                else
+                   snapshots = "#{snapshots},#{m[:ebs_snapshot_id]}"
+                end 
+	     end
+	  end    
        end
+    rescue 
+       return
     end   
-       answer = FXMessageBox.question(@ec2_main.tabBook,MBOX_YES_NO,"Confirm","Confirm Deregister of Image "+sel_image+" and Delete of Snapshot "+snapshot_id)
-       if answer == MBOX_CLICKED_YES
-        if @ec2 != nil
-  	   begin
-              @ec2.deregister_image(sel_image)
-           rescue
+    answer = FXMessageBox.question(@ec2_main.tabBook,MBOX_YES_NO,"Confirm","Confirm Deregister of Image #{sel_image} and Delete of Snapshot(s) #{snapshots}")
+    if answer == MBOX_CLICKED_YES
+       if @ec2 != nil
+          begin
+             @ec2.deregister_image(sel_image)
+          rescue
              error_message("DeRegister of Image failed",$!.to_s)
              return
-           end
-           if snapshot_id != nil and snapshot_id != "" 
-              delete_snapshot(snapshot_id)
-           end   
-           @deleted = true
-        else
-      	   puts "***Error: No EC2 Connection"
-        end
-       end    
+          end
+          snapshot_ids.each {|s| delete_snapshot(s) }
+          @deleted = true
+       else
+      	  puts "***Error: No EC2 Connection"
+       end
+    end    
   end 
    
   def get_image(image_id)
@@ -72,14 +79,16 @@ class EC2_ImageEBSDeleteDialog < FXDialogBox
   end 
   
   def delete_snapshot(snapshot_id)
+    if snapshot_id != nil and snapshot_id != "" 
       ec2 = @ec2_main.environment.connection
       if ec2 != nil
-  	 begin 
+          begin 
             ec2.delete_snapshot(snapshot_id)
          rescue
             error_message("EBS Snapshot Deletion failed",$!.to_s)
          end   
       end
+    end  
   end     
    
    def deleted
